fix: change return type of get_current_supply function from i128 to u32

+++ contracts/minah/src/lib.rs
@@ -574,7 +574,7 @@ impl Minah {
     }
 
     /// Get current supply
-    pub fn get_current_supply(e: Env) -> i128 {
+    pub fn get_current_supply(e: Env) -> u32 {
         e.storage()
             .instance()
             .get(&DataKey::CurrentSupply)
